Maximum-Likelihood Estimation of a Poisson-GP Model

Maximum-Likelihood Estimation of a Poisson-GP model using heterogeneous data.

Usage

# S3 method for poisGP
MLE(
  object = NULL,
  parIni = NULL,
  estim = c("optim", "nloptr", "eval", "none"),
  coefLower,
  coefUpper,
  parTrack = FALSE,
  scale = FALSE,
  trace = 0,
  ...
)

Arguments

object

A poisGP object that needs to be estimated.

parIni

Initial values for the parameter vector. This is must be a named vector of length \(2\) with elements names "scale" and "shape".

estim

Type or method chosen for the estimation.

coefLower, coefUpper

Lower and Upper bounds for the parameters. The should be numeric vectors with names in c("lambda", "scale", "shape"). Only the bounds on the GP parameters "scale" and "shape" can be used during the estimation and they will only be used when estim is "nloptr". However the bounds are used in the inference confint.poisGP and RL.poisGP.

parTrack

Not used yet.

scale

Logical. If TRUE the data used in the optimisation are scaled , see threshData. NOT IMPLEMENTED YET.

trace

Integer Level of verbosity.

...

Not used.

Value

A list with the results of the likelihood maximisation. The content of the list depends on the method as given by estim, yet it should always contain an element

logLik giving the maximised log-likelihood.

Details

The estimation proceeds by minimising a concentrated (or profile) negative log-likelihood which depends on the two GPD parameters, but not on the Poisson rate. So provided bounds on this parameter (if any) will have no effect on the estimation and the estimates can fail to have its element "lambda" within the bounds when these are not 0.0 and Inf. However the standard (non-profile) negative log-likelihood function is built and returned because it will be used to derive profile-likelihood inference results.

Author

Yves Deville