Autoplot Method for poisGP Objects

Autoplot method for objects representing fitted Poisson-GP models.

Usage

# S3 method for poisGP
autoplot(object, 
         which = c("RL", "pp"),
         level = 0.70,
         points = c("H", "p", "none"), a = 0.5,
         allPoints = FALSE,
         trace = 0,
         ...)

Arguments

object

An object with class "poisGP".

which

The type of plot to be built: "RL" is for a Return Level plot and "pp" is for a probability plot.

level

Used when which is "RL" and then passed to RL.poisGP.

points

The plotting position system to be used to show the data attached to the object. When points is "none", the data are not shown. With the value "H" the data are shown using the "H-points" of Nelson's plotting positions. With the value "p", the data are shown using the "p-points" are used and the argument a can be passed to the ppoints function.

a

The parameter used to compute the plotting positions when points is "p". See RP.potData.

allPoints

Logical. With the default FALSE, the limits for the y axis are set to show only the points over the threshold. The value TRUE leads to showing all points in the data including those which are below the threshold.

trace

Integer level of verbosity. Passed to RL.poisGP.

...

Not used.

Value

A ggplot graphics object.

See also