Peaks Over Threshold and/or Block Maxima

Estimation and inference for standard Peak Over Threshold (POT) models using heterogeneous data. The results can be formulated in relation either with the Poisson-GP or with the Point-Process formulation. The classical block maxima and r-largest frameworks can be coped with, and moreover the data can in both cases be censored by discarding the observations below a chosen threshold.

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.

As Renext does this package focuses on data that can be thought of as generated by a marked process called Poisson-GP. Events occur at random Poisson times \(T_i\) and each event comes with a mark \(Y_i\) with Generalised Pareto Distribution (GPD). The marks are assumed to be independent and to be independent of the event times \(T_i\). This process is POT-stable: if the observations are thinned by selecting the marks that are over a threshold \(v\), we get a new Poisson-GP process with smaller rate \(\lambda S(v)\) where \(S(y)\) denotes the survival function of the GPD; the excesses \(Y_i - v\) follow a GPD with its the shape \(\xi\) remaining the same.

The estimation and inference can use heterogenous data where some observations are not necessarily couples \(T_i,\,Y_i]\) and can correspond to time intervals called blocks. For some blocks the complete observations can be available. For some other blocks called MAX we have partial observations: the block maximum or the \(r\) largest order statistics where \(r\) is assumed to be given. For some other blocks called OTS we have different partial observations: the events with mark \(Y_i\) exceeding a block-specific threshold. These two schemes of partial observation can be met when working with historical data, an OTS threshold being then often called a perception threshold.

As it is done in Renext, the ML estimate of the parameter \([\lambda,\, \sigma, \,\xi]\) for the general heterogenous data is obtained by maximising a profile log-likelihood depending on the two GPD parameters \(\sigma\) and \(\xi\) but not on the Poisson rate \(\lambda\). So the optimisation is a two-parameter one.

There are two major new possibilities not found in Renext.

The OTS and MAX data are no longer assumed to have their threshold or data over the "main" threshold, say \(u\). This makes possible to censor block maxima or \(r\)-largest data by keeping only the observations \(> u\).

The preferred method to infer on the parameters and on the return levels is profile likelihood. The determination of the confidence interval is obtained by using a new method which uses only constrained optimisation: the profile-likelihood function is not computed as such and no zero-finding is used. The methods confint and RL have an optional argument check that can be used to check the profile-likelhood results with a classical graphical method as found in other extreme-value packages like ismev or extremes.

Another difference with Renext concerns graphics. While Renext uses classical graphics from the graphics package, ggplot graphics produced with the ggplot2 package are here considered. This means that autoplot and autolayer methods should be used where the classical graphics would be produced with plot and points or lines.

Note that the package does not allow the use of non-stationary Poisson-GP models.

Author

Yves Deville [cre, aut] (<https://orcid.org/0000-0002-1233-488X>)

Maintainer: Yves Deville <deville.yves@alpestat.com>

References

Yves Deville(2020) Renext Computing Details. Tech. Report.

See also